Moved targeting logic to lib file
